+++ temba/channels/handlers.py
@@ -2,30 +2,31 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import json
+import logging
+from datetime import datetime
+
 import pytz
 import six
-import logging
-
-from datetime import datetime
 from django.conf import settings
 from django.db.models import Q
 from django.http import HttpResponse, JsonResponse
 from django.utils.encoding import force_text
 from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import View
+from twilio import twiml
 
 from temba.api.models import WebHookEvent
 from temba.channels.models import Channel, ChannelLog
 from temba.contacts.models import Contact, URN
 from temba.flows.models import Flow, FlowRun, FlowStep
-from temba.orgs.models import NEXMO_UUID
 from temba.msgs.models import Msg, HANDLE_EVENT_TASK, HANDLER_QUEUE, MSG_EVENT
+from temba.orgs.models import NEXMO_UUID
 from temba.triggers.models import Trigger
 from temba.ussd.models import USSDSession
 from temba.utils import on_transaction_commit
-from temba.utils.queues import push_task
 from temba.utils.http import HttpEvent
-from twilio import twiml
+from temba.utils.queues import push_task
+from temba.utils.text import decode_base64
 
 logger = logging.getLogger(__name__)
 
@@ -79,9 +80,6 @@ def get_channel_handlers():
 
 
 class TwimlAPIHandler(BaseChannelHandler):
-
-    courier_url = r'^tw/(?P<uuid>[a-z0-9\-]+)/(?P<action>receive|status)$'
-    courier_name = 'courier.tw'
 
     handler_url = r'^twiml_api/(?P<uuid>[a-z0-9\-]+)/?$'
     handler_name = 'handlers.twiml_api_handler'
@@ -168,6 +166,8 @@ class TwimlAPIHandler(BaseChannelHandler):
                     # either way, we need to hangup now
                     return HttpResponse(six.text_type(response))
 
+        action = request.GET.get('action', 'received')
+
         # check for call progress events, these include post-call hangup notifications
         if request.POST.get('CallbackSource', None) == 'call-progress-events':
             if call_sid:
@@ -179,6 +179,78 @@ class TwimlAPIHandler(BaseChannelHandler):
                     call.save()
                     return HttpResponse("Call status updated")
             return HttpResponse("No call found")
+
+        # this is a callback for a message we sent
+        elif action == 'callback':  # pragma: no cover
+            logger.error("Twilio callback handler called in RapidPro with URL: %s" % request.get_full_path())
+
+            smsId = request.GET.get('id', None)
+            status = request.POST.get('SmsStatus', None)
+
+            # get the SMS
+            sms = Msg.objects.select_related('channel').filter(id=smsId).first()
+            if sms is None:
+                return HttpResponse("No message found with id: %s" % smsId, status=400)
+
+            # validate this request is coming from twilio
+            org = sms.org
+
+            if not org.is_connected_to_twilio():
+                return HttpResponse("No Twilio account is connected", status=400)
+
+            channel = sms.channel
+            if not channel:  # pragma: needs cover
+                channel = org.channels.filter(channel_type=self.get_channel_type()).first()
+
+            client = self.get_client(channel=channel)
+            validator = RequestValidator(client.auth[1])
+
+            if not validator.validate(url, request.POST, signature):  # pragma: needs cover
+                return HttpResponse("Invalid request signature.", status=400)
+
+            # queued, sending, sent, failed, or received.
+            if status == 'sent':
+                sms.status_sent()
+            elif status == 'delivered':
+                sms.status_delivered()
+            elif status == 'failed':
+                sms.status_fail()
+
+            return HttpResponse("", status=200)
+
+        elif action == 'received':  # pragma: no cover
+            logger.error("Twilio receive handler called in RapidPro with URL: %s" % request.get_full_path())
+
+            if not to_number:
+                return HttpResponse("Must provide To number for received messages", status=400)
+
+            channel = self.get_receive_channel(uuid=channel_uuid)
+            if not channel:
+                return HttpResponse("No active channel found for number: %s" % to_number, status=400)
+
+            org = channel.org
+
+            if self.get_channel_type() == 'T' and not org.is_connected_to_twilio():
+                return HttpResponse("No Twilio account is connected", status=400)
+
+            client = self.get_client(channel=channel)
+            validator = RequestValidator(client.auth[1])
+
+            if not validator.validate(url, request.POST, signature):
+                return HttpResponse("Invalid request signature.", status=400)
+
+            # Twilio sometimes sends concat sms as base64 encoded MMS
+            body = decode_base64(request.POST['Body'])
+            urn = URN.from_tel(request.POST['From'])
+            attachments = []
+
+            # download any attached media
+            for i in range(int(request.POST.get('NumMedia', 0))):
+                attachments.append(client.download_media(request.POST['MediaUrl%d' % i]))
+
+            Msg.create_incoming(channel, urn, body, attachments=attachments)
+
+            return HttpResponse("", status=201)
 
         return HttpResponse("Not Handled, unknown action", status=400)  # pragma: no cover
 
@@ -196,9 +268,6 @@ class TwimlAPIHandler(BaseChannelHandler):
 
 
 class TwilioHandler(TwimlAPIHandler):
-
-    courier_url = r'^t/(?P<uuid>[a-z0-9\-]+)/(?P<action>receive|status|voice)$'
-    courier_name = 'courier.t'
 
     handler_url = r'^twilio/(?P<action>receive|status|voice)/(?P<uuid>[a-z0-9\-]+)/?$'
     handler_name = 'handlers.twilio_handler'